refactor(tasks): extract version choices and analysis to prompts/version-choices.ts
